core: add Handler, Looper and HandlerThread.
Provides implementations for Handler, HandlerThread and Looper. The
classes are mostly inspired from the Android Open Source project but
only keeps the most used methods. More will (or not) be added later.

A Looper is a Message orchestrator that uses a MessageQueue to dispatch
messages on their Handler. It allows only one Looper instance per
thread. The Handler is a Message processor. It receives a message and
executes the corresponding action. Its implementation is up to the user.
A HandlerThread provides a easy way to create a thread for a Looper. It
is a kind of utility class.