message queue: add optimized insertion case.
As we use a linked list we have the possibility to easilky check the
last list element. As most of the messages are inserted with the current
time, they are inserted at the end of the list. Thus we can add a check
to see if we can directly insert at the end of the list.